+++ _swaptions/src/nr_routines.cpp
@@ -111,7 +111,7 @@ void gaussj(FTYPE **a, int n, FTYPE **b, int m)
 
 
 /**********************************************************************/
-int *ivector(long nl, long nh)
+int *ivector(int nl, int nh)
 /* allocate an int vector with subscript range v[nl..nh] */
 {
 	int *v;
@@ -122,14 +122,14 @@ int *ivector(long nl, long nh)
 }
 
 /**********************************************************************/
-void free_ivector(int *v, long nl, long nh)
+void free_ivector(int *v, int nl, int nh)
 /* free an int vector allocated with ivector() */
 {
 	free((char *) (v+nl-1));
 }
 
 /**********************************************************************/
-FTYPE *dvector( long nl, long nh )
+FTYPE *dvector( int nl, int nh )
 {
   // allocate a FTYPE vector with subscript range v[nl..nh]
 
@@ -142,7 +142,7 @@ FTYPE *dvector( long nl, long nh )
 } // end of dvector
 
 /**********************************************************************/
-void free_dvector( FTYPE *v, long nl, long nh )
+void free_dvector( FTYPE *v, int nl, int nh )
 {
   // free a FTYPE vector allocated with dvector()
 
@@ -151,11 +151,11 @@ void free_dvector( FTYPE *v, long nl, long nh )
 } // end of free_dvector
 
 /**********************************************************************/
-FTYPE **dmatrix( long nrl, long nrh, long ncl, long nch )
+FTYPE **dmatrix( int nrl, int nrh, int ncl, int nch )
 {
   // allocate a FTYPE matrix with subscript range m[nrl..nrh][ncl..nch]
 
-	long i, nrow=nrh-nrl+1,ncol=nch-ncl+1;
+	int i, nrow=nrh-nrl+1,ncol=nch-ncl+1;
 	FTYPE **m;
 
   // allocate pointers to rows
@@ -178,7 +178,7 @@ FTYPE **dmatrix( long nrl, long nrh, long ncl, long nch )
 } // end of dmatrix 
 
 /**********************************************************************/
-void free_dmatrix( FTYPE **m, long nrl, long nrh, long ncl, long nch )
+void free_dmatrix( FTYPE **m, int nrl, int nrh, int ncl, int nch )
 {
   // free a FTYPE matrix allocated by dmatrix()
